refactor: lightpush options and be more explicit about the relay node being present or not
